requirements.txt
Django==1.9.6
django-bootstrap3==7.0.1
django-filter==0.13.0
django-modeltranslation==0.11
djangorestframework==3.3.3
Markdown==2.6.6
psycopg2==2.6.1